fix: extract pr watcher attach from the panel controller
Move attachWatcher/detachWatcher and the watcher poll gate into a dedicated mixin so the controller stays under the 500-line ratchet without a baseline bump.
